Extracted local search code into separate file.
